feat: add eval of variables from config files
The config files can now use a $ prepended to the value of an argument
to a program to signal that the value should be evaluated. The value
is evaluated in a context which contains the arguments to the program
and a set of helper methods.

The context currently consist of:
 * file - the file command line argument
 * project_root(file) - the first git root in any of the arg parent dir
 * test_file(file) - the test file parallel of the given file
 * advisor_url - the url of a page containing the component