refactor: move i3 implementation details out of the layout
This change does a few things:
1) Moves the i3 defined layout structure into the i3 wm
2) Adds the app into the nodes in the layout structure
3) Removes the app list from the workspace and replace it with a helper
method on the workspace (app_list)
4) Removes the duplicate parsing for apps

This refactoring is done to remove the i3 implementation details from
the general layout and to put the app configuration into the layout
three. This is a preparation step for bringing workflow to platforms
which are not using i3. The more complete layout three can now be used
to calculate and generate the tile layout without a tiling window
manager.