test: exit with return code from tests
The last command in the script for executing jest tests with the xvfb
framebuffer is not the actual test execution. Returning this return
code would produce false positives in the ci execution as it inspects
the return code of the script. This change makes ci-i3 return the exit
code of the `yarn test command`.