test: add echos and retries to i3 ci script
Retries are added when staring the Xvfb and i3 process. This aims to
solve and issue where the integration tests reports connection issues
with the i3 process. echos are added for easier debugging of failed
tests.

In addition we change the killall command to a specific kill -p
<pid>. This will prevent the script from killing unwanted Xvfb
processes. Killing the Xvfb process will kill the i3 process so this
is not killed explicitly for simplicity.