feat: implement configuration render using react
the new package react-workflow introduces a react renderer which takes
a react component and translates it into a workflow configuration.

The package introduces the three lowlevel components Workflow, Layout
and App which will be used to build more abstract components for
rendering complex layouts and Applications.

A proper example is included in the demo/App.js. The main difference
in syntax between this example and the hacked React.createElement
examples in the current version of the workflow packages are that the
example only depends on classes in react-workflow and that the default
export needs to be rendered using the render function exported by
react-workflow.