fix: generate workflow-home with versions from devDependencies
The devDependencies in the package.json from the root of the
create-workflow-home package are automatically handled by the release process.
This makes these variables much easier to maintain then the hardcoded variables
in the old versions file. This commit makes the create-workflow-home package get
versions from the packages json instead.