feat: add yargs command line optstring parser
Adds yargs as a library for optstring parser. This provides a simple way of
adding new commands to the `workflow` cli and good support for `help` message.

The default implementation will read the `config.js` file and `flow` file to
provide details for arguments given in the approporiate flow. The arguments for
plugins like `resolvers`, `wms` etc. will be added as a part of #89.

Fixes #15