chore(package.json): add yarn build to release
This forces the build to be done before releasing with `lerna publish`.
The alternative would be to have `prepublish` scripts in each package
to rely on the `yarn publish` command to run them. This could be a more
efficient approach as it would only build the require package. However we opt
for the safer an more naive verison first.